requirements.txt
selenium==4.21.0
requests
python-telegram-bot==13.15
python-dotenv
pytz